refine unittest, add dynamic convert test
